move opposed matrix tests to use a shared base class
